move all jalousie tests into single dir
