+++ src/modules/jalousie/test/ConfigTest.cpp
